+++ rust/rust-codegen/src/main.rs
@@ -61,6 +61,7 @@ fn main() -> color_eyre::Result<()> {
         proto_dir.join("astarteplatform/msghub/astarte_type.proto"),
         proto_dir.join("astarteplatform/msghub/config.proto"),
         proto_dir.join("astarteplatform/msghub/interface.proto"),
+        proto_dir.join("astarteplatform/msghub/property.proto"),
     ];
 
     // NOTE: This is a temporary workaround to build the documentation on docs.rs, since they are